updated test and cache
